Migrate additional test
